fixed some bugs on request and preprocess
